Updated sensor_math(initial commit)

+++ include/sensor_math.hpp
@@ -1,8 +1,13 @@
 #ifndef SENSOR_MATH_HPP
 #define SENSOR_MATH_HPP
 #include <Arduino.h>
+
+
 inline void initializeSensorMath();
-inline void applyIMUCalibration(int16_t& ax, int16_t& ay, int16_t& az, int16_t& gx, int16_t& gy, int16_t& gz, int16_t& mx, int16_t& my, int16_t& mz);
-inline void applyDepthSensorCalibration(int16_t& depth);
-inline void updateOrientation(int16_t ax, int16_t ay, int16_t az, int16_t gx, int16_t gy, int16_t gz, int16_t mx, int16_t my, int16_t mz, float& roll, float& pitch, float& yaw);
+inline void applyIMUCalibration(float &ax, float &ay, float &az, float &gx, float &gy, float &gz, float &mx, float &my, float &mz);
+inline void applyDepthSensorCalibration(float &depth);
+inline void updateOrientation(float ax, float ay, float az, float gx, float gy, float gz, float mx, float my, float mz, float &roll, float &pitch, float &yaw);
+
 #endif // SENSOR_MATH_HPP
+
+
