+++ src/main.cpp
@@ -1,6 +1,7 @@
 #include "main.hpp"
-
-int16_t ax, ay, az, gx, gy, gz, mx, my, mz, depth;
+#include "communication_interface.hpp"
+#include "Fusion/Fusion.h"
+float ax, ay, az, gx, gy, gz, mx, my, mz, depth;
 float roll, pitch, yaw;
 int16_t current_mode = 0;
 void setup() {
